// src/types/onboarding.ts
export enum ActionType {
  CHANGE_SITE = "CHANGE_SITE",
  NAME = "NAME",
  SURNAME = "SURNAME",
  PROFILEIMAGE = "PROFILEIMAGE",
  USECASE = "USECASE",
  WORKSPACE_NAME = "WORKSPACE_NAME",
  WORKSPACE_IMAGE = "WORKSPACE_IMAGE",
}

export enum UseCase {
  WORK = "WORK",
  STUDY = "STUDY",
  PERSONAL_USE = "PERSONAL_USE",
}

export interface Action {
  type: ActionType;
  payload: string | number | UseCase | undefined | null;
}

export interface OnboardingFormReducer {
  currentStep: 1 | 2 | 3 | 4 | 5;
  name?: string | null;
  profileImage?: string | null;
  useCase: UseCase | null;
  workspaceName: string | null;
  workspaceImage?: string | null;
}

export interface OnboardingFormContext extends OnboardingFormReducer {
  dispatch: React.Dispatch<Action>;
}